Fixed Sabertooth watchdog
Watchdog wasn't enabling, since the bogie controller was booting up faster than
the sabertooths, and sending the watchdog message before the bogies were ready
to accept them.  The fix was adding a 100ms delay.  It's not the best fix, but
there's nothing in the datasheet about this, and there's no indication from the
sabertooth that we can look for to tell that it's working (other than the status
LEDs).