Added code for digital potentiometers
They're actually called digital magnetic encoders, but since I get an
absolute position out of them (like a continuous-turn potentiometer),
I'm calling them digital potentiometers.
Anyway, the code uses SPI, and it seems to work. What we have to watch
out for, however, is the shaft slipping.  It has plenty of ability to
do so, and there's no limit switches for me to index off of.  However,
as long as the shaft doesn't slip, the limits appear very repeatable,
and the mechanism shouldn't break itself apart.

= Marshal