Arm recalibrated
Should be within 1 degree, ignoring mechanical slop.
Also added limits to keep from commanding the arm out of range.
The pv calculation was changed so the offset is subtracted before the
value is scaled.
The only other thing I can think of is possibly limiting the power to
the motors if the current gets too high.

- Marshal