Encoders aren't returning reliable data
Or maybe it's the PID manipulation.  Anyway, the values printed back
often don't match up with the value I told it to stop at.
Maybe it's a problem with get_turn() being a macro?
Maybe I need to make the encoder counting purely event-system?
Maybe TCC1.CNT is not volatile, and the reads/writes are being
optimized?  This would certainly cause trouble.